fix: updated code as per code review
