+++ app/client/src/PluginActionEditor/components/PluginActionResponse/components/Schema/CurrentDataSource.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Flex, Link } from "@appsmith/ads";
+import { Flex } from "@appsmith/ads";
 import { getAssetUrl } from "ee/utils/airgapHelpers";
 import { EntityIcon } from "pages/Editor/Explorer/ExplorerIcons";
 import { useSelector } from "react-redux";
@@ -7,41 +7,27 @@ import {
   getPluginIdFromDatasourceId,
   getPluginImages,
 } from "ee/selectors/entitiesSelector";
-import { useDataSourceNavigation } from "ee/PluginActionEditor/hooks/useDataSourceNavigation";
 
-interface iProps {
-  type: "link" | "trigger";
+interface Props {
   datasourceId: string;
   datasourceName: string;
 }
 
-const CurrentDataSource = ({ datasourceId, datasourceName, type }: iProps) => {
+const CurrentDataSource = ({ datasourceId, datasourceName }: Props) => {
   const { pluginId, pluginImages } = useSelector((state) => ({
     pluginId: getPluginIdFromDatasourceId(state, datasourceId),
     pluginImages: getPluginImages(state),
   }));
 
-  const { goToDatasource } = useDataSourceNavigation();
-
   const datasourceIcon = pluginId ? pluginImages?.[pluginId] : undefined;
 
-  const content = (
-    <Flex
-      alignItems="center"
-      gap="spaces-2"
-      justifyContent={type === "link" ? "center" : "start"}
-    >
+  return (
+    <Flex alignItems="center" gap="spaces-2">
       <EntityIcon height="16px" width="16px">
         <img alt="entityIcon" src={getAssetUrl(datasourceIcon)} />
       </EntityIcon>
       {datasourceName}
     </Flex>
-  );
-
-  return type === "link" ? (
-    <Link onClick={() => goToDatasource(datasourceId)}>{content}</Link>
-  ) : (
-    content
   );
 };
