+++ app/client/src/PluginActionEditor/components/PluginActionResponse/components/Schema/constants.tsx
@@ -1,0 +1,2 @@
+// This is bottom bar height + paddings
+export const BOTTOMBAR_HEIGHT = 42;
