--- app/client/src/ee/PluginActionEditor/hooks/useDataSourceNavigation.ts
@@ -1,1 +1,0 @@
-export * from "ce/PluginActionEditor/hooks/useDataSourceNavigation";
