+++ app/client/src/ee/PluginActionEditor/hooks/useGoToDatasource.ts
@@ -1,0 +1,1 @@
+export * from "ce/PluginActionEditor/hooks/useGoToDatasource";
